requirements.txt
BeautifulSoup==3.2.1
Django==1.4.3
chardet==2.1.1
clint==0.3.1
cssselect==0.7.1
gensim==0.8.6
lxml==3.1beta1
networkx==1.7
numpy==1.6.2
readability-lxml==0.2.6.1
scikit-learn==0.13
scipy==0.11.0
sunlight==1.1.7
wsgiref==0.1.2
